move `collectParentsAndSiblings` to `MetadataUtils`
